+++ Directory.Build.props
@@ -118,7 +118,7 @@ A basic lambda function can handle it - the only reason this function is as comp
     <TestSdkVersion>17.11.1</TestSdkVersion>
     <CoverletVersion>6.0.3</CoverletVersion>
      <XunitRunneVisualstudio>3.1.4</XunitRunneVisualstudio>
-    <AkkaVersion>1.5.51-alpha-1759266431</AkkaVersion>
+    <AkkaVersion>1.5.51</AkkaVersion>
     <MicrosoftExtensionsVersion>[6.0.0,)</MicrosoftExtensionsVersion>
     <SystemTextJsonVersion>[6.0.10,)</SystemTextJsonVersion>
   </PropertyGroup>
