+++ nuget.config
@@ -4,13 +4,8 @@
     <!--To inherit the global NuGet package sources remove the <clear/> line below -->
     <clear />
     <add key="nuget.org" value="https://api.nuget.org/v3/index.json" />
-    <add key="testlab" value="https://nuget.testlab.petabridge.net/v3/index.json"/>
   </packageSources>
   <packageSourceMapping>
-    <packageSource key="testlab">
-      <package pattern="Akka"/>
-      <package pattern="Akka.*"/>
-    </packageSource>
     <packageSource key="nuget.org">
       <package pattern="*" />
     </packageSource>
